Split off TreeBuilder
